requirements.txt
django==5.0.6
djangorestframework==3.15.2
psycopg==3.2.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
alpha_vantage==2.3.1
pycoingecko==3.1.0